refactoring the integration test into separate test class for each integration
